In Progress

+++ TestSutraUsers/src/main/java/com/tathvatech/injuryReport/oid/InjuryOID.java
@@ -1,9 +1,10 @@
 package com.tathvatech.injuryReport.oid;
 
-import com.tathvatech.ts.core.common.EntityTypeEnum;
-import com.tathvatech.ts.core.common.OID;
 
-public class InjuryOID extends OID{
+import com.tathvatech.common.enums.EntityTypeEnum;
+import com.tathvatech.user.OID.OID;
+
+public class InjuryOID extends OID {
 
     public InjuryOID(int pk, String displayText)
     {
@@ -32,7 +33,7 @@ public class InjuryOID extends OID{
 
     @Override
     public int hashCode() {
-        return super.getPk();
+        return (int) super.getPk();
     }
 
 }
